Add specific add_action for genesis_header_right
It appears Genesis has code to skip the genesis_header_right action if
no widgets appear in this widget area.  In which case, the hook does not
display.

To avoid this problem, we are hooking into genesis_header_right and
displaying nothing (i.e. an empty string '').  Even though we're
displaying an empty string, this is enough to allow the
genesis_header_right action to occur.

I've not spotted any other actions that behave in this manner but if
there are, we'll need to add them manually, as well.

Specifically, neither the Primary nor Secondary sidebar show this
behavior.

See #25
See #24